Add util.js module
